test(faxsms): cover reminder runner orchestrator and DTO
Adds isolated test coverage for the new appointment-reminder pipeline so
the patch ships above codecov's 50% target instead of relying solely on
the original 16 helper tests.

- ReminderRunResultTest pins the snake_case toArray() shape that PSR-3
  logger context arrays depend on across modules.
- AppointmentNotificationRunnerTest gains 13 orchestrator tests that
  exercise run() through fake AppDispatch/EmailClient subclasses (empty
  constructors skip the heavy session/crypto setup) plus a TestableRunner
  that stubs fetchPendingAppointments. Stubs for the procedural log/mark
  helpers live in a separate file using bracketed namespace syntax so
  they can publish symbols in the runner's namespace and the global
  spy class in one shot.
- Marks the legacy procedural entry points (rc_sms_notification.php,
  rc_sms_notification_helpers.php, run_notifications.php) and the
  BackgroundReminderTask bridge as @codeCoverageIgnore. They each pull in
  live module bootstrap state (AppDispatch::getApiService, raw \$_GET
  writes, globals.php require) that an isolated test cannot supply; the
  scan-and-send pipeline below them is covered in the runner test.
- Drops the unnecessary (int) casts on NotificationTaskManager::getTaskHours()
  in rc_sms_notification.php and BackgroundReminderTask.php (RecastingRemovalRector
  flagged them as no-ops).
- Fixes a PHP 8.5 implicit-nullable deprecation on
  AppDispatch::getSession(\$param) so loading the class in isolated tests
  no longer pollutes output and break the session-storage test suites
  that run later.

+++ interface/modules/custom_modules/oe-module-faxsms/src/Notification/BackgroundReminderTask.php
@@ -25,6 +25,15 @@
  * @author    Michael A. Smith <michael@opencoreemr.com>
  * @copyright Copyright (c) 2026 OpenCoreEMR Inc <https://opencoreemr.com/>
  * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
+ *
+ * @codeCoverageIgnore Thin bridge: resolves the channel client through
+ *     `AppDispatch::getApiService()` (DB-backed), narrows credentials,
+ *     and constructs the runner. Each step pulls in live module
+ *     bootstrap state that an isolated test cannot supply. The
+ *     scan-and-send pipeline below this bridge is covered in
+ *     `AppointmentNotificationRunnerTest`. The follow-up issue #11848
+ *     proposes replacing this class with a PSR-11-resolved job, at
+ *     which point the bridge itself goes away.
  */
 
 declare(strict_types=1);
@@ -81,7 +90,7 @@ final class BackgroundReminderTask
         $vendor = AppDispatch::getModuleVendor();
         $gatewayType = is_string($vendor) ? $vendor : '';
 
-        $cronIntervalHours = (int) (new NotificationTaskManager())->getTaskHours(strtolower($channel->value));
+        $cronIntervalHours = (new NotificationTaskManager())->getTaskHours(strtolower($channel->value));
 
         // The background path never runs in dry-run mode; live delivery
         // is the whole point of the scheduled task. Keep bTestRun in
